Split up campaigns.Store in multiple files
While adding `ChangesetSpecs` and `CampaignSpecs` I noticed how much
easier it is to work with the store and its tests when you have
different files for different tables.

So, this PR does _nothing_ besides splitting up the code that was in
`store.go` into `store.go`, `store_campaigns.go`,
`store_changeset_events.go` and `store_changesets.go`.
