+++ enterprise/internal/campaigns/store_campaigns.go
@@ -1,0 +1,388 @@
+package campaigns
+
+import (
+	"context"
+
+	"github.com/keegancsmith/sqlf"
+	"github.com/sourcegraph/sourcegraph/internal/campaigns"
+	"github.com/sourcegraph/sourcegraph/internal/db/dbutil"
+)
+
+// CreateCampaign creates the given Campaign.
+func (s *Store) CreateCampaign(ctx context.Context, c *campaigns.Campaign) error {
+	q, err := s.createCampaignQuery(c)
+	if err != nil {
+		return err
+	}
+
+	return s.query(ctx, q, func(sc scanner) (err error) {
+		return scanCampaign(c, sc)
+	})
+}
+
+var createCampaignQueryFmtstr = `
+-- source: enterprise/internal/campaigns/store.go:CreateCampaign
+INSERT INTO campaigns (
+  name,
+  description,
+  branch,
+  author_id,
+  namespace_user_id,
+  namespace_org_id,
+  created_at,
+  updated_at,
+  changeset_ids,
+  closed_at,
+  campaign_spec_id
+)
+VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s)
+RETURNING
+  id,
+  name,
+  description,
+  branch,
+  author_id,
+  namespace_user_id,
+  namespace_org_id,
+  created_at,
+  updated_at,
+  changeset_ids,
+  closed_at,
+  campaign_spec_id
+`
+
+func (s *Store) createCampaignQuery(c *campaigns.Campaign) (*sqlf.Query, error) {
+	changesetIDs, err := jsonSetColumn(c.ChangesetIDs)
+	if err != nil {
+		return nil, err
+	}
+
+	if c.CreatedAt.IsZero() {
+		c.CreatedAt = s.now()
+	}
+
+	if c.UpdatedAt.IsZero() {
+		c.UpdatedAt = c.CreatedAt
+	}
+
+	return sqlf.Sprintf(
+		createCampaignQueryFmtstr,
+		c.Name,
+		c.Description,
+		c.Branch,
+		c.AuthorID,
+		nullInt32Column(c.NamespaceUserID),
+		nullInt32Column(c.NamespaceOrgID),
+		c.CreatedAt,
+		c.UpdatedAt,
+		changesetIDs,
+		nullTimeColumn(c.ClosedAt),
+		nullInt64Column(c.CampaignSpecID),
+	), nil
+}
+
+// UpdateCampaign updates the given Campaign.
+func (s *Store) UpdateCampaign(ctx context.Context, c *campaigns.Campaign) error {
+	q, err := s.updateCampaignQuery(c)
+	if err != nil {
+		return err
+	}
+
+	return s.query(ctx, q, func(sc scanner) (err error) { return scanCampaign(c, sc) })
+}
+
+var updateCampaignQueryFmtstr = `
+-- source: enterprise/internal/campaigns/store.go:UpdateCampaign
+UPDATE campaigns
+SET (
+  name,
+  description,
+  branch,
+  author_id,
+  namespace_user_id,
+  namespace_org_id,
+  updated_at,
+  changeset_ids,
+  closed_at,
+  campaign_spec_id
+) = (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s)
+WHERE id = %s
+RETURNING
+  id,
+  name,
+  description,
+  branch,
+  author_id,
+  namespace_user_id,
+  namespace_org_id,
+  created_at,
+  updated_at,
+  changeset_ids,
+  closed_at,
+  campaign_spec_id
+`
+
+func (s *Store) updateCampaignQuery(c *campaigns.Campaign) (*sqlf.Query, error) {
+	changesetIDs, err := jsonSetColumn(c.ChangesetIDs)
+	if err != nil {
+		return nil, err
+	}
+
+	c.UpdatedAt = s.now()
+
+	return sqlf.Sprintf(
+		updateCampaignQueryFmtstr,
+		c.Name,
+		c.Description,
+		c.Branch,
+		c.AuthorID,
+		nullInt32Column(c.NamespaceUserID),
+		nullInt32Column(c.NamespaceOrgID),
+		c.UpdatedAt,
+		changesetIDs,
+		nullTimeColumn(c.ClosedAt),
+		nullInt64Column(c.CampaignSpecID),
+		c.ID,
+	), nil
+}
+
+// DeleteCampaign deletes the Campaign with the given ID.
+func (s *Store) DeleteCampaign(ctx context.Context, id int64) error {
+	return s.Store.Exec(ctx, sqlf.Sprintf(deleteCampaignQueryFmtstr, id))
+}
+
+var deleteCampaignQueryFmtstr = `
+-- source: enterprise/internal/campaigns/store.go:DeleteCampaign
+DELETE FROM campaigns WHERE id = %s
+`
+
+// CountCampaignsOpts captures the query options needed for
+// counting campaigns.
+type CountCampaignsOpts struct {
+	ChangesetID int64
+	State       campaigns.CampaignState
+	// Only return campaigns where author_id is the given.
+	OnlyForAuthor int32
+}
+
+// CountCampaigns returns the number of campaigns in the database.
+func (s *Store) CountCampaigns(ctx context.Context, opts CountCampaignsOpts) (int, error) {
+	return s.queryCount(ctx, countCampaignsQuery(&opts))
+}
+
+var countCampaignsQueryFmtstr = `
+-- source: enterprise/internal/campaigns/store.go:CountCampaigns
+SELECT COUNT(id)
+FROM campaigns
+WHERE %s
+`
+
+func countCampaignsQuery(opts *CountCampaignsOpts) *sqlf.Query {
+	var preds []*sqlf.Query
+	if opts.ChangesetID != 0 {
+		preds = append(preds, sqlf.Sprintf("changeset_ids ? %s", opts.ChangesetID))
+	}
+
+	switch opts.State {
+	case campaigns.CampaignStateOpen:
+		preds = append(preds, sqlf.Sprintf("closed_at IS NULL"))
+	case campaigns.CampaignStateClosed:
+		preds = append(preds, sqlf.Sprintf("closed_at IS NOT NULL"))
+	}
+
+	if opts.OnlyForAuthor != 0 {
+		preds = append(preds, sqlf.Sprintf("author_id = %d", opts.OnlyForAuthor))
+	}
+
+	if len(preds) == 0 {
+		preds = append(preds, sqlf.Sprintf("TRUE"))
+	}
+
+	return sqlf.Sprintf(countCampaignsQueryFmtstr, sqlf.Join(preds, "\n AND "))
+}
+
+// GetCampaignOpts captures the query options needed for getting a Campaign
+type GetCampaignOpts struct {
+	ID int64
+
+	NamespaceUserID int32
+	NamespaceOrgID  int32
+
+	CampaignSpecID int64
+	Name           string
+}
+
+// GetCampaign gets a campaign matching the given options.
+func (s *Store) GetCampaign(ctx context.Context, opts GetCampaignOpts) (*campaigns.Campaign, error) {
+	q := getCampaignQuery(&opts)
+
+	var c campaigns.Campaign
+	err := s.query(ctx, q, func(sc scanner) error {
+		return scanCampaign(&c, sc)
+	})
+	if err != nil {
+		return nil, err
+	}
+
+	if c.ID == 0 {
+		return nil, ErrNoResults
+	}
+
+	return &c, nil
+}
+
+var getCampaignsQueryFmtstr = `
+-- source: enterprise/internal/campaigns/store.go:GetCampaign
+SELECT
+  campaigns.id,
+  campaigns.name,
+  campaigns.description,
+  campaigns.branch,
+  campaigns.author_id,
+  campaigns.namespace_user_id,
+  campaigns.namespace_org_id,
+  campaigns.created_at,
+  campaigns.updated_at,
+  campaigns.changeset_ids,
+  campaigns.closed_at,
+  campaigns.campaign_spec_id
+FROM campaigns
+WHERE %s
+LIMIT 1
+`
+
+func getCampaignQuery(opts *GetCampaignOpts) *sqlf.Query {
+	var preds []*sqlf.Query
+	if opts.ID != 0 {
+		preds = append(preds, sqlf.Sprintf("campaigns.id = %s", opts.ID))
+	}
+
+	if opts.CampaignSpecID != 0 {
+		preds = append(preds, sqlf.Sprintf("campaigns.campaign_spec_id = %s", opts.CampaignSpecID))
+	}
+
+	if opts.NamespaceUserID != 0 {
+		preds = append(preds, sqlf.Sprintf("campaigns.namespace_user_id = %s", opts.NamespaceUserID))
+	}
+
+	if opts.NamespaceOrgID != 0 {
+		preds = append(preds, sqlf.Sprintf("campaigns.namespace_org_id = %s", opts.NamespaceOrgID))
+	}
+
+	if opts.Name != "" {
+		preds = append(preds, sqlf.Sprintf("campaigns.name = %s", opts.Name))
+
+	}
+
+	if len(preds) == 0 {
+		preds = append(preds, sqlf.Sprintf("TRUE"))
+	}
+
+	return sqlf.Sprintf(
+		getCampaignsQueryFmtstr,
+		sqlf.Join(preds, "\n AND "),
+	)
+}
+
+// ListCampaignsOpts captures the query options needed for
+// listing campaigns.
+type ListCampaignsOpts struct {
+	ChangesetID int64
+	Cursor      int64
+	Limit       int
+	State       campaigns.CampaignState
+	// Only return campaigns where author_id is the given.
+	OnlyForAuthor int32
+}
+
+// ListCampaigns lists Campaigns with the given filters.
+func (s *Store) ListCampaigns(ctx context.Context, opts ListCampaignsOpts) (cs []*campaigns.Campaign, next int64, err error) {
+	q := listCampaignsQuery(&opts)
+
+	cs = make([]*campaigns.Campaign, 0, opts.Limit)
+	err = s.query(ctx, q, func(sc scanner) error {
+		var c campaigns.Campaign
+		if err := scanCampaign(&c, sc); err != nil {
+			return err
+		}
+		cs = append(cs, &c)
+		return nil
+	})
+
+	if opts.Limit != 0 && len(cs) == opts.Limit {
+		next = cs[len(cs)-1].ID
+		cs = cs[:len(cs)-1]
+	}
+
+	return cs, next, err
+}
+
+var listCampaignsQueryFmtstr = `
+-- source: enterprise/internal/campaigns/store.go:ListCampaigns
+SELECT
+  id,
+  name,
+  description,
+  branch,
+  author_id,
+  namespace_user_id,
+  namespace_org_id,
+  created_at,
+  updated_at,
+  changeset_ids,
+  closed_at,
+  campaign_spec_id
+FROM campaigns
+WHERE %s
+ORDER BY id ASC
+LIMIT %s
+`
+
+func listCampaignsQuery(opts *ListCampaignsOpts) *sqlf.Query {
+	if opts.Limit == 0 {
+		opts.Limit = defaultListLimit
+	}
+	opts.Limit++
+
+	preds := []*sqlf.Query{
+		sqlf.Sprintf("id >= %s", opts.Cursor),
+	}
+
+	if opts.ChangesetID != 0 {
+		preds = append(preds, sqlf.Sprintf("changeset_ids ? %s", opts.ChangesetID))
+	}
+
+	switch opts.State {
+	case campaigns.CampaignStateOpen:
+		preds = append(preds, sqlf.Sprintf("closed_at IS NULL"))
+	case campaigns.CampaignStateClosed:
+		preds = append(preds, sqlf.Sprintf("closed_at IS NOT NULL"))
+	}
+
+	if opts.OnlyForAuthor != 0 {
+		preds = append(preds, sqlf.Sprintf("author_id = %d", opts.OnlyForAuthor))
+	}
+
+	return sqlf.Sprintf(
+		listCampaignsQueryFmtstr,
+		sqlf.Join(preds, "\n AND "),
+		opts.Limit,
+	)
+}
+
+func scanCampaign(c *campaigns.Campaign, s scanner) error {
+	return s.Scan(
+		&c.ID,
+		&c.Name,
+		&dbutil.NullString{S: &c.Description},
+		&dbutil.NullString{S: &c.Branch},
+		&c.AuthorID,
+		&dbutil.NullInt32{N: &c.NamespaceUserID},
+		&dbutil.NullInt32{N: &c.NamespaceOrgID},
+		&c.CreatedAt,
+		&c.UpdatedAt,
+		&dbutil.JSONInt64Set{Set: &c.ChangesetIDs},
+		&dbutil.NullTime{Time: &c.ClosedAt},
+		&dbutil.NullInt64{N: &c.CampaignSpecID},
+	)
+}
