+++ enterprise/internal/campaigns/store_changeset_events.go
@@ -1,0 +1,341 @@
+package campaigns
+
+import (
+	"context"
+	"encoding/json"
+	"fmt"
+	"time"
+
+	"github.com/keegancsmith/sqlf"
+	"github.com/pkg/errors"
+	"github.com/sourcegraph/sourcegraph/internal/campaigns"
+)
+
+// GetChangesetEventOpts captures the query options needed for getting a ChangesetEvent
+type GetChangesetEventOpts struct {
+	ID          int64
+	ChangesetID int64
+	Kind        campaigns.ChangesetEventKind
+	Key         string
+}
+
+// GetChangesetEvent gets a changeset matching the given options.
+func (s *Store) GetChangesetEvent(ctx context.Context, opts GetChangesetEventOpts) (*campaigns.ChangesetEvent, error) {
+	q := getChangesetEventQuery(&opts)
+
+	var c campaigns.ChangesetEvent
+	err := s.query(ctx, q, func(sc scanner) error {
+		return scanChangesetEvent(&c, sc)
+	})
+	if err != nil {
+		return nil, err
+	}
+
+	if c.ID == 0 {
+		return nil, ErrNoResults
+	}
+
+	return &c, nil
+}
+
+var getChangesetEventsQueryFmtstr = `
+-- source: enterprise/internal/campaigns/store.go:GetChangesetEvent
+SELECT
+    id,
+    changeset_id,
+    kind,
+    key,
+    created_at,
+    updated_at,
+    metadata
+FROM changeset_events
+WHERE %s
+LIMIT 1
+`
+
+func getChangesetEventQuery(opts *GetChangesetEventOpts) *sqlf.Query {
+	var preds []*sqlf.Query
+	if opts.ID != 0 {
+		preds = append(preds, sqlf.Sprintf("id = %s", opts.ID))
+	}
+
+	if opts.ChangesetID != 0 && opts.Kind != "" && opts.Key != "" {
+		preds = append(preds,
+			sqlf.Sprintf("changeset_id = %s", opts.ChangesetID),
+			sqlf.Sprintf("kind = %s", opts.Kind),
+			sqlf.Sprintf("key = %s", opts.Key),
+		)
+	}
+
+	if len(preds) == 0 {
+		preds = append(preds, sqlf.Sprintf("TRUE"))
+	}
+
+	return sqlf.Sprintf(getChangesetEventsQueryFmtstr, sqlf.Join(preds, "\n AND "))
+}
+
+// ListChangesetEventsOpts captures the query options needed for
+// listing changeset events.
+type ListChangesetEventsOpts struct {
+	ChangesetIDs []int64
+	Cursor       int64
+	Limit        int
+}
+
+// ListChangesetEvents lists ChangesetEvents with the given filters.
+func (s *Store) ListChangesetEvents(ctx context.Context, opts ListChangesetEventsOpts) (cs []*campaigns.ChangesetEvent, next int64, err error) {
+	q := listChangesetEventsQuery(&opts)
+
+	cs = make([]*campaigns.ChangesetEvent, 0, opts.Limit)
+	err = s.query(ctx, q, func(sc scanner) (err error) {
+		var c campaigns.ChangesetEvent
+		if err = scanChangesetEvent(&c, sc); err != nil {
+			return err
+		}
+		cs = append(cs, &c)
+		return nil
+	})
+
+	if opts.Limit != 0 && len(cs) == opts.Limit {
+		next = cs[len(cs)-1].ID
+		cs = cs[:len(cs)-1]
+	}
+
+	return cs, next, err
+}
+
+var listChangesetEventsQueryFmtstr = `
+-- source: enterprise/internal/campaigns/store.go:ListChangesetEvents
+SELECT
+    id,
+    changeset_id,
+    kind,
+    key,
+    created_at,
+    updated_at,
+    metadata
+FROM changeset_events
+WHERE %s
+ORDER BY id ASC
+`
+
+func listChangesetEventsQuery(opts *ListChangesetEventsOpts) *sqlf.Query {
+	if opts.Limit == 0 {
+		opts.Limit = defaultListLimit
+	}
+	opts.Limit++
+
+	var limitClause string
+	if opts.Limit > 0 {
+		limitClause = fmt.Sprintf("LIMIT %d", opts.Limit)
+	}
+
+	preds := []*sqlf.Query{
+		sqlf.Sprintf("id >= %s", opts.Cursor),
+	}
+
+	if len(opts.ChangesetIDs) != 0 {
+		ids := make([]*sqlf.Query, 0, len(opts.ChangesetIDs))
+		for _, id := range opts.ChangesetIDs {
+			if id != 0 {
+				ids = append(ids, sqlf.Sprintf("%d", id))
+			}
+		}
+		preds = append(preds,
+			sqlf.Sprintf("changeset_id IN (%s)", sqlf.Join(ids, ",")))
+	}
+
+	return sqlf.Sprintf(
+		listChangesetEventsQueryFmtstr+limitClause,
+		sqlf.Join(preds, "\n AND "),
+	)
+}
+
+// CountChangesetEventsOpts captures the query options needed for
+// counting changeset events.
+type CountChangesetEventsOpts struct {
+	ChangesetID int64
+}
+
+// CountChangesetEvents returns the number of changeset events in the database.
+func (s *Store) CountChangesetEvents(ctx context.Context, opts CountChangesetEventsOpts) (int, error) {
+	return s.queryCount(ctx, countChangesetEventsQuery(&opts))
+}
+
+var countChangesetEventsQueryFmtstr = `
+-- source: enterprise/internal/campaigns/store.go:CountChangesetEvents
+SELECT COUNT(id)
+FROM changeset_events
+WHERE %s
+`
+
+func countChangesetEventsQuery(opts *CountChangesetEventsOpts) *sqlf.Query {
+	var preds []*sqlf.Query
+	if opts.ChangesetID != 0 {
+		preds = append(preds, sqlf.Sprintf("changeset_id = %s", opts.ChangesetID))
+	}
+
+	if len(preds) == 0 {
+		preds = append(preds, sqlf.Sprintf("TRUE"))
+	}
+
+	return sqlf.Sprintf(countChangesetEventsQueryFmtstr, sqlf.Join(preds, "\n AND "))
+}
+
+// UpsertChangesetEvents creates or updates the given ChangesetEvents.
+func (s *Store) UpsertChangesetEvents(ctx context.Context, cs ...*campaigns.ChangesetEvent) (err error) {
+	q, err := s.upsertChangesetEventsQuery(cs)
+	if err != nil {
+		return err
+	}
+
+	i := -1
+	return s.query(ctx, q, func(sc scanner) (err error) {
+		i++
+		return scanChangesetEvent(cs[i], sc)
+	})
+}
+
+const changesetEventsBatchQueryPrefix = `
+WITH batch AS (
+  SELECT * FROM ROWS FROM (
+  json_to_recordset(%s)
+  AS (
+      id           bigint,
+      changeset_id integer,
+      kind         text,
+      key          text,
+      created_at   timestamptz,
+      updated_at   timestamptz,
+      metadata     jsonb
+    )
+  )
+  WITH ORDINALITY
+)
+`
+
+const batchChangesetEventsQuerySuffix = `
+SELECT
+  changed.id,
+  changed.changeset_id,
+  changed.kind,
+  changed.key,
+  changed.created_at,
+  changed.updated_at,
+  changed.metadata
+FROM changed
+LEFT JOIN batch
+ON batch.changeset_id = changed.changeset_id
+AND batch.kind = changed.kind
+AND batch.key = changed.key
+ORDER BY batch.ordinality
+`
+
+var upsertChangesetEventsQueryFmtstr = changesetEventsBatchQueryPrefix + `,
+-- source: enterprise/internal/campaigns/store.go:UpsertChangesetEvents
+changed AS (
+  INSERT INTO changeset_events (
+    changeset_id,
+    kind,
+    key,
+    created_at,
+    updated_at,
+    metadata
+  )
+  SELECT
+    changeset_id,
+    kind,
+    key,
+    created_at,
+    updated_at,
+    metadata
+  FROM batch
+  ON CONFLICT ON CONSTRAINT
+    changeset_events_changeset_id_kind_key_unique
+  DO UPDATE
+  SET
+    metadata   = excluded.metadata,
+    updated_at = excluded.updated_at
+  RETURNING changeset_events.*
+)
+` + batchChangesetEventsQuerySuffix
+
+func (s *Store) upsertChangesetEventsQuery(es []*campaigns.ChangesetEvent) (*sqlf.Query, error) {
+	now := s.now()
+	for _, e := range es {
+		if e.CreatedAt.IsZero() {
+			e.CreatedAt = now
+		}
+
+		if !e.UpdatedAt.After(e.CreatedAt) {
+			e.UpdatedAt = now
+		}
+	}
+	return batchChangesetEventsQuery(upsertChangesetEventsQueryFmtstr, es)
+}
+
+func batchChangesetEventsQuery(fmtstr string, es []*campaigns.ChangesetEvent) (*sqlf.Query, error) {
+	type record struct {
+		ID          int64           `json:"id"`
+		ChangesetID int64           `json:"changeset_id"`
+		Kind        string          `json:"kind"`
+		Key         string          `json:"key"`
+		CreatedAt   time.Time       `json:"created_at"`
+		UpdatedAt   time.Time       `json:"updated_at"`
+		Metadata    json.RawMessage `json:"metadata"`
+	}
+
+	records := make([]record, 0, len(es))
+
+	for _, e := range es {
+		metadata, err := jsonbColumn(e.Metadata)
+		if err != nil {
+			return nil, err
+		}
+
+		records = append(records, record{
+			ID:          e.ID,
+			ChangesetID: e.ChangesetID,
+			Kind:        string(e.Kind),
+			Key:         e.Key,
+			CreatedAt:   e.CreatedAt,
+			UpdatedAt:   e.UpdatedAt,
+			Metadata:    metadata,
+		})
+	}
+
+	batch, err := json.MarshalIndent(records, "    ", "    ")
+	if err != nil {
+		return nil, err
+	}
+
+	return sqlf.Sprintf(fmtstr, string(batch)), nil
+}
+
+func scanChangesetEvent(e *campaigns.ChangesetEvent, s scanner) error {
+	var metadata json.RawMessage
+
+	err := s.Scan(
+		&e.ID,
+		&e.ChangesetID,
+		&e.Kind,
+		&e.Key,
+		&e.CreatedAt,
+		&e.UpdatedAt,
+		&metadata,
+	)
+	if err != nil {
+		return err
+	}
+
+	e.Metadata, err = campaigns.NewChangesetEventMetadata(e.Kind)
+	if err != nil {
+		return err
+	}
+
+	if err = json.Unmarshal(metadata, e.Metadata); err != nil {
+		return errors.Wrapf(err, "scanChangesetEvent: failed to unmarshal %q metadata", e.Kind)
+	}
+
+	return nil
+}
